+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJCMEGM0F3H3D4MNJGHW6TG7"
+        "01KJCSPZTE564ZTZ87GG5HTAPP"
     )
     r"""The unique identifier of the data source"""